Fix stop app scriptwq

+++ scripts/stop_service.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
 
 isExistApp=`pgrep java`
-kill -9 $isExistApp  
+if [[ -z $isExistApp ]]
+then
+     echo "Application is not running"
+     exit 1
+else
+    echo "Application pid $isExistApp stopping"
+    kill -9 $isExistApp
+fi
